+++ src/ch/idsia/ai/agents/AmiCoAgent.java
@@ -23,7 +23,7 @@ public class AmiCoAgent implements Agent
         this.reset();
     }
 
-    public void integrateObservation(byte[] serializedLevelSceneObservationZ, byte[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
+    public void integrateObservation(int[] serializedLevelSceneObservationZ, int[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
     {
         //To change body of implemented methods use File | Settings | File Templates.
     }
