+++ src/ch/idsia/ai/agents/SimpleAgent.java
@@ -15,7 +15,7 @@ public class SimpleAgent implements Agent
     protected boolean Action[] = new boolean[Environment.numberOfButtons];
     protected String Name = "SimpleAgent";
 
-    public void integrateObservation(byte[] serializedLevelSceneObservationZ, byte[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
+    public void integrateObservation(int[] serializedLevelSceneObservationZ, int[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
     {
         //To change body of implemented methods use File | Settings | File Templates.
     }
